fix visibility issues for tabs panels and sidebars
