+++ app/components/avo/item_switcher_component.html.erb
@@ -6,7 +6,7 @@
   <%= render Avo::PanelComponent.new(name: item.name, description: item.description, index: index) do |c| %>
     <% c.body do %>
       <div class="divide-y">
-        <% item.items.each_with_index do |field, index| %>
+        <% item.visible_items.each_with_index do |field, index| %>
           <%= render field
             .hydrate(resource: @resource, model: @resource.model, user: resource.user, view: view)
             .component_for_view(view)
