+++ lib/avo/panel.rb
@@ -1,16 +1,19 @@
 class Avo::Panel
   include Avo::Concerns::IsResourceItem
+  include Avo::Concerns::VisibleItems
 
   class_attribute :item_type, default: :panel
 
   attr_reader :name
+  attr_reader :view
   attr_reader :description
   attr_accessor :items_holder
 
   delegate :items, :add_item, to: :items_holder
 
-  def initialize(name: nil, description: nil)
+  def initialize(name: nil, description: nil, view: nil)
     @name = name
+    @view = view
     @description = description
     @items_holder = Avo::ItemsHolder.new
   end
@@ -21,5 +24,11 @@ class Avo::Panel
 
   def has_items?
     @items.present?
+  end
+
+  def hydrate(view: nil)
+    @view = view
+
+    self
   end
 end
