+++ lib/avo/sidebar.rb
@@ -1,5 +1,6 @@
 class Avo::Sidebar
   include Avo::Concerns::IsResourceItem
+  include Avo::Concerns::VisibleItems
   include Avo::Fields::FieldExtensions::VisibleInDifferentViews
 
   class_attribute :item_type, default: :sidebar
@@ -31,30 +32,5 @@ class Avo::Sidebar
 
   def empty?
     visible_items.blank?
-  end
-
-  def items
-    if self.items_holder.present?
-      self.items_holder.items
-    else
-      []
-    end
-  end
-
-  def visible_items
-    items.map do |item|
-      # Remove the fields that shouldn't be visible in this view
-      # eg: has_many fields on edit
-      if not_visible_field(item)
-        nil
-      else
-        item
-      end
-    end
-    .compact
-  end
-
-  def not_visible_field(item)
-    item.is_field? && !item.visible_on?(view)
   end
 end
